+++ Tests/KituraSampleRouterTests/KituraTest.swift
@@ -92,18 +92,23 @@ class KituraTest: XCTestCase {
         }
         req.end()
     }
-
-    func performRequestSynchronous(_ method: String, path: String,  expectation: XCTestExpectation,
+    
+    func performRequest(_ method: String,
+                        path: String,
+                        body: String,
+                        expectation: XCTestExpectation,
                         headers: [String: String]? = nil,
-                        requestModifier: ((ClientRequest) -> Void)? = nil,
-                        callback: @escaping (ClientResponse, DispatchGroup) -> Void) {
-        let dispatchGroup = DispatchGroup()
-        dispatchGroup.enter()
-        performRequest(method, path: path, expectation: expectation, headers: headers,
-                       requestModifier: requestModifier) { response in
-                        callback(response, dispatchGroup)
+                        callback: @escaping (ClientResponse) -> Void) {
+        self.performRequest(method,
+                            path: path,
+                            expectation: expectation,
+                            headers: headers,
+                            requestModifier: { request in
+                                request.write(from: body)
+                            }
+        ) { response in
+            callback(response)
         }
-        dispatchGroup.wait()
     }
 
     func expectation(_ index: Int) -> XCTestExpectation {
@@ -113,5 +118,138 @@ class KituraTest: XCTestCase {
 
     func waitExpectation(timeout t: TimeInterval, handler: XCWaitCompletionHandler?) {
         self.waitForExpectations(timeout: t, handler: handler)
+    }
+    
+    func runTestThatCorrectHTMLTitleIsReturned(expectedTitle: String, path: String) {
+        let pattern = "<title>(.*?)</title>"
+        
+        runGetResponseTest(path: path) { body in
+            do {
+                #if os(Linux) && !swift(>=3.1)
+                    let regularExpressionOptional: RegularExpression? =
+                    try RegularExpression(pattern: pattern, options: [])
+                #else
+                    let regularExpressionOptional: NSRegularExpression? =
+                        try NSRegularExpression(pattern: pattern, options: [])
+                #endif
+                guard let regularExpression = regularExpressionOptional else {
+                    XCTFail("failed to create regular expression")
+                    return
+                }
+                
+                let matches = regularExpression.matches(in: body, options: [],
+                                                        range: NSMakeRange(0, body.count))
+                
+                guard let match = matches.first else {
+                    XCTFail("no match of title tag in body")
+                    return
+                }
+                
+                let titleRange = match.range(at: 1)
+                let titleInBody = NSString(string: body).substring(with: titleRange)
+                XCTAssertEqual(titleInBody, expectedTitle,
+                               "returned title does not match the expected one")
+            } catch {
+                XCTFail("failed to create regular expression: \(error)")
+            }
+        }
+    }
+    
+    typealias BodyChecker =  (String) -> Void
+    func checkResponse(response: ClientResponse,
+                       expectedResponseText: String? = nil,
+                       expectedStatusCode: HTTPStatusCode = HTTPStatusCode.OK,
+                       bodyChecker: BodyChecker? = nil) {
+        XCTAssertEqual(response.statusCode, expectedStatusCode, "No success status code returned")
+        if let optionalBody = try? response.readString(), let body = optionalBody {
+            if let expectedResponseText = expectedResponseText {
+                XCTAssertEqual(body, expectedResponseText, "mismatch in body")
+            }
+            bodyChecker?(body)
+        } else {
+            XCTFail("No response body")
+        }
+    }
+    
+    func runGetResponseTest(path: String, expectedResponseText: String? = nil,
+                            expectedStatusCode: HTTPStatusCode = HTTPStatusCode.OK,
+                            bodyChecker: BodyChecker? = nil) {
+        performServerTest { expectation in
+            self.performRequest("get", path: path, expectation: expectation) { response in
+                self.checkResponse(response: response, expectedResponseText: expectedResponseText,
+                                   expectedStatusCode: expectedStatusCode, bodyChecker: bodyChecker)
+                expectation.fulfill()
+            }
+        }
+    }
+    
+    func runTestUser(expectedUser: String, expectation: XCTestExpectation) {
+        self.performRequest("get", path: "/hello", expectation: expectation) {
+            response in
+            self.checkResponse(response: response,
+                               expectedResponseText: "Hello \(expectedUser), from Kitura!")
+            expectation.fulfill()
+        }
+    }
+
+    
+    func runTestModifyUser(method: String, userToSet: String? = nil,
+                           expectation: XCTestExpectation) {
+        self.performRequest(method, path: "/hello", expectation: expectation,
+                            requestModifier: { request in
+                                if let userToSet = userToSet {
+                                    request.write(from: userToSet)
+                                }
+        }) { response in
+            self.checkResponse(response: response,
+                               expectedResponseText: "Got a \(method.uppercased()) request")
+            expectation.fulfill()
+        }
+    }
+    
+    func runGetCodableResponseTest(path: String, expectedResponse: Book? = nil, expectedStatusCode: HTTPStatusCode = HTTPStatusCode.OK) {
+        performServerTest { expectation in
+            self.performRequest("get", path: path, expectation: expectation) { response in
+                self.checkCodableResponse(response: response, expectedResponse: expectedResponse,
+                                          expectedStatusCode: expectedStatusCode)
+                expectation.fulfill()
+            }
+        }
+    }
+    
+    func checkCodableResponse(response: ClientResponse, expectedResponse: Book? = nil,
+                              expectedStatusCode: HTTPStatusCode = HTTPStatusCode.OK) {
+        XCTAssertEqual(response.statusCode, expectedStatusCode,
+                       "No success status code returned")
+        if let optionalBody = try? response.readString(), let body = optionalBody {
+            var responseString = body
+            if(responseString.hasPrefix("[")) {
+                responseString.removeFirst()
+                responseString.removeLast()
+            }
+            if responseString == "" {
+                XCTAssertNil(expectedResponse)
+            } else {
+                let json = responseString.data(using: .utf8)!
+                do {
+                    let myStruct = try JSONDecoder().decode(Book.self, from: json)
+                    if let expectedBook = expectedResponse {
+                        XCTAssertEqual(myStruct.name, expectedBook.name)
+                        XCTAssertEqual(myStruct.author, expectedBook.author)
+                        XCTAssertEqual(myStruct.rating, expectedBook.rating)
+                    }
+                } catch {
+                    print("Error")
+                }
+            }
+        } else {
+            XCTFail("No response body")
+        }
+    }
+    
+    func runTestParameter(user: String) {
+        let userInPath = user.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed) ?? user
+        let responseText = "<!DOCTYPE html><html><body><b>User:</b> \(user)</body></html>\n\n"
+        runGetResponseTest(path: "/users/\(userInPath)", expectedResponseText: responseText)
     }
 }
